Complete overhaul of slideshow
Rather than cycling through an array of images in JS the images
are all placed in the HTML. JS changes their class names so that
one is in the class 'showing' at all times where opacity: 1 and the rest
have an opacity of 0. This makes it easy to fade between them because
we can just use CSS transition on the property opacity, rather than
doing it with JS or JQuery.
Based off of:
https://www.sitepoint.com
/make-a-simple-javascript-slideshow-without-jquery/