Add new pictures (still not perfect)
added overflow attribute to slideshow container with value of
hidden so that when the window is resized it doesn't try to cram the
picture in a small space. I would like to get it like the ray johnson
website where it zooms out allowing the picture to fit in a smaller window
without crunching it.